requirements.txt
playwright==1.58.0
openai
python-dotenv
